Reward extra accurate minutia matches

+++ SourceAFIS/Matching/Minutia/MinutiaPair.cs
@@ -6,13 +6,13 @@ namespace SourceAFIS.Matching.Minutia
 {
     public struct MinutiaPair
     {
-        public int Probe;
-        public int Candidate;
+        public short Probe;
+        public short Candidate;
 
         public MinutiaPair(int probe, int candidate)
         {
-            Probe = probe;
-            Candidate = candidate;
+            Probe = (short)probe;
+            Candidate = (short)candidate;
         }
     }
 }
